+++ nn_dataflow/tests/dataflow_test/test_nn_dataflow.py
@@ -54,6 +54,7 @@ class TestNNDataflow(unittest.TestCase):
                                  size_gbuf=128 * 1024 // 2,  # 128 kB
                                  size_regf=512 // 2,  # 512 B
                                  array_bus_width=float('inf'),
+                                 dram_bandwidth=float('inf'),
                                 )
 
         self.cost = Cost(mac_op=1,
@@ -140,6 +141,7 @@ class TestNNDataflow(unittest.TestCase):
                                  size_gbuf=128 * 1024 // 2,  # 128 kB
                                  size_regf=2,
                                  array_bus_width=float('inf'),
+                                 dram_bandwidth=float('inf'),
                                 )
 
         nnd = NNDataflow(self.alex_net, 4, self.resource, self.cost,
@@ -257,6 +259,7 @@ class TestNNDataflow(unittest.TestCase):
                             size_gbuf=108 * 1024 // 2,  # 108 kB
                             size_regf=261,  # 225 + 12 + 24
                             array_bus_width=float('inf'),
+                            dram_bandwidth=float('inf'),
                            )
 
         cost = Cost(mac_op=2e-12,
@@ -346,6 +349,7 @@ class TestNNDataflow(unittest.TestCase):
                             size_gbuf=576056 // 2,  # 576 kB
                             size_regf=1024 // 2,  # 1 kB
                             array_bus_width=float('inf'),
+                            dram_bandwidth=float('inf'),
                            )
 
         cost = Cost(mac_op=2e-12,
@@ -371,6 +375,7 @@ class TestNNDataflow(unittest.TestCase):
                             size_gbuf=133032 // 2,  # 133 kB
                             size_regf=512 // 2,  # 512 B
                             array_bus_width=float('inf'),
+                            dram_bandwidth=float('inf'),
                            )
 
         cost = Cost(mac_op=2e-12,