+++ nn_dataflow/tests/unit_test/test_resource.py
@@ -42,12 +42,14 @@ class TestResource(unittest.TestCase):
                             size_gbuf=131072,
                             size_regf=512,
                             array_bus_width=8,
+                            dram_bandwidth=128,
                            )
         self.assertTupleEqual(resource.proc_region.dim, (2, 2), 'proc_region')
         self.assertTupleEqual(resource.dim_array, (16, 16), 'dim_array')
         self.assertEqual(resource.size_gbuf, 131072, 'size_gbuf')
         self.assertEqual(resource.size_regf, 512, 'size_regf')
         self.assertEqual(resource.array_bus_width, 8, 'array_bus_width')
+        self.assertEqual(resource.dram_bandwidth, 128, 'dram_bandwidth')
 
     def test_invalid_proc_region(self):
         ''' Invalid proc_region. '''
@@ -63,6 +65,7 @@ class TestResource(unittest.TestCase):
                          size_gbuf=131072,
                          size_regf=512,
                          array_bus_width=8,
+                         dram_bandwidth=128,
                         )
 
     def test_invalid_proc_region_data(self):
@@ -81,6 +84,7 @@ class TestResource(unittest.TestCase):
                          size_gbuf=131072,
                          size_regf=512,
                          array_bus_width=8,
+                         dram_bandwidth=128,
                         )
 
     def test_invalid_dim_array(self):
@@ -99,6 +103,7 @@ class TestResource(unittest.TestCase):
                          size_gbuf=131072,
                          size_regf=512,
                          array_bus_width=8,
+                         dram_bandwidth=128,
                         )
 
     def test_invalid_size_gbuf(self):
@@ -117,6 +122,7 @@ class TestResource(unittest.TestCase):
                          size_gbuf=(131072,),
                          size_regf=512,
                          array_bus_width=8,
+                         dram_bandwidth=128,
                         )
 
     def test_invalid_size_regf(self):
@@ -135,6 +141,7 @@ class TestResource(unittest.TestCase):
                          size_gbuf=131072,
                          size_regf=(512,),
                          array_bus_width=8,
+                         dram_bandwidth=128,
                         )
 
     def test_invalid_array_bus_width(self):
@@ -154,6 +161,7 @@ class TestResource(unittest.TestCase):
                          size_gbuf=131072,
                          size_regf=512,
                          array_bus_width=1.2,
+                         dram_bandwidth=128,
                         )
         with self.assertRaisesRegexp(ValueError,
                                      'Resource: .*array_bus_width.*'):
@@ -170,6 +178,7 @@ class TestResource(unittest.TestCase):
                          size_gbuf=131072,
                          size_regf=512,
                          array_bus_width=-2,
+                         dram_bandwidth=128,
                         )
         with self.assertRaisesRegexp(ValueError,
                                      'Resource: .*array_bus_width.*'):
@@ -186,6 +195,61 @@ class TestResource(unittest.TestCase):
                          size_gbuf=131072,
                          size_regf=512,
                          array_bus_width=0,
+                         dram_bandwidth=128,
+                        )
+
+    def test_invalid_dram_bandwidth(self):
+        ''' Invalid dram_bandwidth. '''
+        with self.assertRaisesRegexp(TypeError,
+                                     'Resource: .*dram_bandwidth.*'):
+            _ = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0),
+                                                type=NodeRegion.PROC),
+                         data_regions=(NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 0),
+                                                  type=NodeRegion.DATA),
+                                       NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 1),
+                                                  type=NodeRegion.DATA)),
+                         dim_array=PhyDim2(16, 16),
+                         size_gbuf=131072,
+                         size_regf=512,
+                         array_bus_width=8,
+                         dram_bandwidth=None,
+                        )
+        with self.assertRaisesRegexp(ValueError,
+                                     'Resource: .*dram_bandwidth.*'):
+            _ = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0),
+                                                type=NodeRegion.PROC),
+                         data_regions=(NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 0),
+                                                  type=NodeRegion.DATA),
+                                       NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 1),
+                                                  type=NodeRegion.DATA)),
+                         dim_array=PhyDim2(16, 16),
+                         size_gbuf=131072,
+                         size_regf=512,
+                         array_bus_width=8,
+                         dram_bandwidth=-3,
+                        )
+        with self.assertRaisesRegexp(ValueError,
+                                     'Resource: .*dram_bandwidth.*'):
+            _ = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0),
+                                                type=NodeRegion.PROC),
+                         data_regions=(NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 0),
+                                                  type=NodeRegion.DATA),
+                                       NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 1),
+                                                  type=NodeRegion.DATA)),
+                         dim_array=PhyDim2(16, 16),
+                         size_gbuf=131072,
+                         size_regf=512,
+                         array_bus_width=8,
+                         dram_bandwidth=0,
                         )
 
     def test_invalid_data_regions_type(self):
@@ -201,6 +265,7 @@ class TestResource(unittest.TestCase):
                          size_gbuf=131072,
                          size_regf=(512,),
                          array_bus_width=8,
+                         dram_bandwidth=128,
                         )
         with self.assertRaisesRegexp(TypeError, 'Resource: .*data_regions.*'):
             _ = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
@@ -213,6 +278,7 @@ class TestResource(unittest.TestCase):
                          size_gbuf=131072,
                          size_regf=(512,),
                          array_bus_width=8,
+                         dram_bandwidth=128,
                         )
         with self.assertRaisesRegexp(TypeError, 'Resource: .*data_regions.*'):
             _ = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
@@ -226,6 +292,7 @@ class TestResource(unittest.TestCase):
                          size_gbuf=131072,
                          size_regf=(512,),
                          array_bus_width=8,
+                         dram_bandwidth=128,
                         )
 
     def test_invalid_data_regions_len(self):
@@ -241,6 +308,7 @@ class TestResource(unittest.TestCase):
                          size_gbuf=131072,
                          size_regf=(512,),
                          array_bus_width=8,
+                         dram_bandwidth=128,
                         )
 
     def test_invalid_data_regions_proc(self):
@@ -259,6 +327,7 @@ class TestResource(unittest.TestCase):
                          size_gbuf=131072,
                          size_regf=512,
                          array_bus_width=8,
+                         dram_bandwidth=128,
                         )
 
     def test_src_data_region(self):
@@ -274,6 +343,7 @@ class TestResource(unittest.TestCase):
                             size_gbuf=131072,
                             size_regf=512,
                             array_bus_width=8,
+                            dram_bandwidth=128,
                             data_regions=(nr1, nr2))
         self.assertEqual(resource.src_data_region(), nr1, 'src_data_region')
         resource = Resource(proc_region=NodeRegion(dim=PhyDim2(4, 4),
@@ -283,6 +353,7 @@ class TestResource(unittest.TestCase):
                             size_gbuf=131072,
                             size_regf=512,
                             array_bus_width=8,
+                            dram_bandwidth=128,
                             data_regions=(nr1,))
         self.assertEqual(resource.src_data_region(), nr1, 'src_data_region')
 
@@ -299,6 +370,7 @@ class TestResource(unittest.TestCase):
                             size_gbuf=131072,
                             size_regf=512,
                             array_bus_width=8,
+                            dram_bandwidth=128,
                             data_regions=(nr1, nr2))
         self.assertEqual(resource.dst_data_region(), nr2, 'dst_data_region')
         resource = Resource(proc_region=NodeRegion(dim=PhyDim2(4, 4),
@@ -308,6 +380,7 @@ class TestResource(unittest.TestCase):
                             size_gbuf=131072,
                             size_regf=512,
                             array_bus_width=8,
+                            dram_bandwidth=128,
                             data_regions=(nr1,))
         self.assertEqual(resource.dst_data_region(), nr1, 'dst_data_region')